+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [0.1.2](https://github.com/indexdata/foliogo/tree/v0.1.2) (IN PROGRESS)
 
 * Elements of `RequestParams` structure are now capitalised, for access from other packages.
+* New method `session.Fetch0` simply invokes `Fetch` with an empty `RequestParams` structure.
 
 ## [0.1.1](https://github.com/indexdata/foliogo/tree/v0.1.1) (2023-11-21)